Spring 5.6 support
