+++ jwt-server/spring/jwt-spring-core/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.entur.jwt.spring.JwtAutoConfiguration
+org.entur.jwt.spring.JwtAutoConfiguration,org.entur.jwt.spring.JwtWebSecurityConfigurerAdapterAutoConfiguration
