+++ pom.xml
@@ -54,29 +54,29 @@
 		<maven.compiler.source>1.8</maven.compiler.source>
 		<maven.compiler.target>1.8</maven.compiler.target>
 
-		<commons-io.version>2.8.0</commons-io.version>
-		<google-truth.version>1.1.2</google-truth.version>
-		<junit.version>5.7.2</junit.version>
-		<mockito.version>3.12.2</mockito.version>
+		<commons-io.version>2.11.0</commons-io.version>
+		<google-truth.version>1.1.3</google-truth.version>
+		<junit.version>5.8.1</junit.version>
+		<mockito.version>4.0.0</mockito.version>
 		<hamcrest.version>2.2</hamcrest.version>
 		<slf4j.version>1.7.32</slf4j.version>
 		<jjwt.version>0.11.2</jjwt.version>
-		<nimbus-jose-jwt.version>9.13</nimbus-jose-jwt.version>
+		<nimbus-jose-jwt.version>9.15.2</nimbus-jose-jwt.version>
 		<bucket4j-core.version>4.10.0</bucket4j-core.version>
-		<jackson-databind.version>2.11.4</jackson-databind.version>
+		<jackson-databind.version>2.13.0</jackson-databind.version>
 		<camel.version>2.25.4</camel.version>
-		<jwks-rsa.version>0.18.0</jwks-rsa.version>
-		<java-jwt.version>3.16.0</java-jwt.version>
+		<jwks-rsa.version>0.20.0</jwks-rsa.version>
+		<java-jwt.version>3.18.2</java-jwt.version>
 		<jadler.version>1.3.0</jadler.version>
-		<spring-boot.version>2.5.4</spring-boot.version>
-		<spring.version>5.3.9</spring.version>
+		<spring-boot.version>2.6.0</spring-boot.version>
+		<spring.version>5.3.13</spring.version>
 		<rest-assured.version>4.3.3</rest-assured.version>
-		  <okhttp.version>4.9.1</okhttp.version>
+		<okhttp.version>4.9.1</okhttp.version>
 
 		<!-- https://repo1.maven.org/maven2/org/springframework/boot/spring-boot-dependencies  -->
 		<!-- transient dependencies added due to security patching -->
-		<snakeyaml.version>1.27</snakeyaml.version>
-		<tomcat.version>9.0.46</tomcat.version>
+		<snakeyaml.version>1.29</snakeyaml.version>
+		<tomcat.version>9.0.55</tomcat.version>
 
 		<!-- plugins -->
 		<build-helper-maven-plugin.version>1.9.1</build-helper-maven-plugin.version>
